test/scripts/javascript-test262-benchmark/requirements.txt
esprima==4.0.1
PyYAML==5.1
